requirements.txt
librato-metrics==0.4.11
docopt==0.6.2
PyYAML==3.11
